# Improved `buildTree`.