--- a/editor/src/benchmarks.benchmark.ts
+++ b/editor/src/benchmarks.benchmark.ts
@@ -5,9 +5,9 @@ import { benchmarkOptics } from './core/shared/optics.benchmark'
 import { benchmarkPropertyPathFunction } from './core/shared/property-path.benchmark'
 import { benchmarkGetUniqueUids } from './core/model/get-unique-ids.benchmark'
 
-// await benchmarkBuildTree()
+await benchmarkBuildTree()
 // await benchmarkElementPathFunction()
 // await benchmarkPropertyPathFunction()
 // await benchmarkAttributes()
 // await benchmarkOptics()
-await benchmarkGetUniqueUids()
+// await benchmarkGetUniqueUids()