--- a/editor/src/components/canvas/controls/select-mode/scene-label.tsx
+++ b/editor/src/components/canvas/controls/select-mode/scene-label.tsx
@@ -18,7 +18,7 @@ import { boundingArea, createInteractionViaMouse } from '../../canvas-strategies
 import { windowToCanvasCoordinates } from '../../dom-lookup'
 import { CanvasOffsetWrapper } from '../canvas-offset-wrapper'
 import { isCommentMode, isSelectModeWithArea } from '../../../editor/editor-modes'
-import { getSubTree } from '../../../../core/shared/element-path-tree'
+import { getElementPathTreeChildren, getSubTree } from '../../../../core/shared/element-path-tree'
 
 interface SceneLabelControlProps {
   maybeHighlightOnHover: (target: ElementPath) => void
@@ -65,7 +65,7 @@ const SceneLabel = React.memo<SceneLabelProps>((props) => {
       if (subTree == null) {
         return false
       } else {
-        return subTree.children.length === 1
+        return getElementPathTreeChildren(subTree).length === 1
       }
     },
     'SceneLabel sceneHasSingleChild',