… editor/src/components/canvas/dom-sampler.spec.browser2.tsx
… editor/src/components/navigator/navigator.spec.browser2.tsx
… editor/src/core/model/element-metadata-utils.spec.tsx
… editor/src/core/shared/element-path-tree.spec.browser2.tsx
… editor/src/core/shared/element-path.ts